Update AutoFish.java with "Safe Fishing" Feature
Im trying to add a new function to AutoFish Module, which makes it pause whenever an ingame screen is open, like inventory, chat or anything else that would not allow you to fish without a mod... This is my first contributon to Java Minecraft so please review it carefully as at the moment this is untested. Have a Good Day!